backend/pyproject.toml

[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "python-dotenv==1.0.1",
    "uvicorn==0.34.0",
    "pydantic==2.10.6",
    "python-multipart==0.0.20",
    "langchain==0.3.21",
    "langchain-openai==0.3.9",
    "beanie==1.29.0",
    "langgraph==0.3.27",
    "pipecat-ai[cartesia,openai,silero,websocket,deepgram,elevenlabs, groq, websocket, langchain]==0.0.65",
    "langchain-mcp-adapters>=0.0.3",
    "mcp>=1.6.0",

]

[dependency-groups]
dev = [
    "pyinstaller>=6.12.0",
    "ruff>=0.8.3",
]
